SFINT-4641 getInsightAnalyticsAction functions added to utils
