+++ packages/headless/src/controllers/insight/facets/facet/headless-insight-facet.ts
@@ -37,14 +37,11 @@ import {loadReducerError} from '../../../../utils/errors';
 import {InsightEngine} from '../../../../app/insight-engine/insight-engine';
 import {
   logFacetClearAll,
-  logFacetDeselect,
-  logFacetSelect,
   logFacetShowLess,
   logFacetShowMore,
   logFacetUpdateSort,
 } from '../../../../features/facets/facet-set/facet-set-insight-analytics-actions';
-import {FacetSelectionChangeMetadata} from '../../../../features/facets/facet-set/facet-set-analytics-actions-utils';
-import {isFacetValueSelected} from '../../../../features/facets/facet-set/facet-set-utils';
+import {getInsightAnalyticsActionForToggleFacetSelect} from '../../../../features/facets/facet-set/facet-set-utils';
 
 export type {
   FacetOptions,
@@ -127,7 +124,7 @@ export function buildFacet(
       coreController.toggleSelect(selection);
       dispatch(
         executeSearch(
-          getAnalyticsActionForToggleFacetSelect(getFacetId(), selection)
+          getInsightAnalyticsActionForToggleFacetSelect(getFacetId(), selection)
         )
       );
     },
@@ -176,18 +173,4 @@ function loadFacetReducers(
 > {
   engine.addReducers({facetSet, configuration, facetSearchSet, search});
   return true;
-}
-
-function getAnalyticsActionForToggleFacetSelect(
-  facetId: string,
-  selection: FacetValue
-) {
-  const payload: FacetSelectionChangeMetadata = {
-    facetId: facetId,
-    facetValue: selection.value,
-  };
-
-  return isFacetValueSelected(selection)
-    ? logFacetDeselect(payload)
-    : logFacetSelect(payload);
 }
