+++ packages/headless/src/features/facets/facet-set/facet-set-utils.ts
@@ -1,5 +1,9 @@
 import {FacetValue} from './interfaces/response';
 import {logFacetDeselect, logFacetSelect} from './facet-set-analytics-actions';
+import {
+  logFacetDeselect as logInsightFacetDeselect,
+  logFacetSelect as logInsightFacetSelect,
+} from './facet-set-insight-analytics-actions';
 import {FacetSelectionChangeMetadata} from './facet-set-analytics-actions-utils';
 
 export const isFacetValueSelected = (value: FacetValue) => {
@@ -18,4 +22,18 @@ export const getAnalyticsActionForToggleFacetSelect = (
   return isFacetValueSelected(selection)
     ? logFacetDeselect(payload)
     : logFacetSelect(payload);
+};
+
+export const getInsightAnalyticsActionForToggleFacetSelect = (
+  facetId: string,
+  selection: FacetValue
+) => {
+  const payload: FacetSelectionChangeMetadata = {
+    facetId: facetId,
+    facetValue: selection.value,
+  };
+
+  return isFacetValueSelected(selection)
+    ? logInsightFacetDeselect(payload)
+    : logInsightFacetSelect(payload);
 };
